Fix DOC201, PLR2004, and RUF067 instead of ignoring
Document return values in all docstrings, replace magic numbers with
named constants and HTTPStatus.OK, and move the implementation to a
private _update_checker module so __init__.py contains only re-exports.
The only remaining global ignores are formatter conflicts and the
copyright notice rule.
